Remove unneeded filter check
The filter checking logic for on_record was incorrect. This just removes
the incorrect check. We'll have checked the filter on `on_new_span`, and
won't find an `idl::Trace` to attach extra data to on the referenced
span. If it was emitted, we'll find and append new annotations.

Testing:

This wasn't tested at all previously, so added a very basic test suite
to src/lib.rs. It sets up a writer that just writes the trace data to a
buffer so that we can verify it after the trace as ended.

The tests check for basic span, and adding arguments with and without
the filter.